% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw the evolution of the number of occurence of a given name}
\usage{
draw_a_name(a_name)
}
\arguments{
\item{a_name}{the name you want to observe}
}
\value{
a graph
}
\description{
Draw the evolution of the number of occurence of a given name
}
\examples{
\dontrun{
draw_a_name("Priscille")
}


}
