% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw the evolution of the occurences of different names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{the names you want to observe}
}
\value{
a ggplot graph
}
\description{
Draw the evolution of the occurences of different names
}
\examples{
\dontrun{
draw_names(c("Jean","Marie"))
}
}
