% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_cvs_reader}
\alias{my_cvs_reader}
\title{import multiple csv files}
\usage{
my_cvs_reader(folder)
}
\arguments{
\item{folder}{path to the folder}
}
\value{
a list
}
\description{
import multiple csv files
}
\examples{
\dontrun{
my_cvs_reader(../../path/to/folder)
}

}
