% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_generale.R
\docType{package}
\name{pack2-package}
\alias{pack2-package}
\alias{pack2}
\title{Deal with excel and csv files and with the prenoms dataset}
\description{
The package has two main objectives :
-deal with excel and csv files (how to extrat all the sheets of an excel with the multi_excel function or how to read a csv file with the csv_reader function)
-deal with the dataset prenoms (draw the evolution of one or more names over the time)
}
\author{
Priscille Boissonnet
}
