% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save as csv function}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{the dataset you work with}

\item{filename}{the file you want to read}

\item{row.names}{the list of the names of the row}
}
\value{
a dataframe
}
\description{
Save as csv function
}
\examples{
\dontrun{
save_as_csv(dataset_name, myfile_with_a_csv_extension)
}
}
